// src/machine-learning/Types.ts
/** Wraps useful information about a data set */
export interface DataSet {
    name:string,
    yLabel:string,
    symbol:string,
    tsData: TimeSeriesData[],
    predictions: Prediction[],
    min:number,
    max:number
}

/** Association between timestamp and a value */
export interface TimeSeriesData {
    timeStamp:number,
    value:number
}

/** Prediction about the future value of a time series
 *  expressed as a probability distribution. */
export interface Prediction{
    timeStamp:number,
    mean: number,
    standardDeviation: number
}

/** The compressibility of a particular data set */
export interface Compressibility {
    name:string,
    compressibility:number
}

/** A sliding window splits data into sequences followed by the next
 *  value. */
export interface SequenceDataItem {
    input:number[],
    nextValue:number,
}

/** TFJS epoch log structure.
 *  Used to handle events emitted while training. */
export interface EpochLog{
    val_loss: number,
    val_acc: number,
    loss: number,
    acc: number
}

/** TFJS batch log structure.
 *  Used to handle events emitted while training. */
export interface BatchLog{
    batch: number,
    size: number,
    loss: number,
    acc: number
}

/** Event generated by a Plotly graph */
export interface PlotlyEvent{
    points:Array<{x:number}>
}